Restructure admin and access-checker
