+++ src/client/app/access-checker/access-checker.module.ts
@@ -4,10 +4,11 @@ import { FormsModule } from '@angular/forms';
 
 import { Ng2BootstrapModule } from 'ng2-bootstrap';
 
+import { UtilModule } from 'app/shared';
+
 import { ViewCacheEntryModal } from './view-cache-entry.component';
 import { CacheEntriesService } from './cache-entries.service';
 import { AdminCacheEntriesComponent } from './admin-cache-entries.component';
-import { UtilModule } from 'app/shared/util.module';
 
 @NgModule({
 	imports: [
