+++ src/client/app/access-checker/cache-entries.service.ts
@@ -3,8 +3,7 @@ import { Response } from '@angular/http';
 
 import { Observable } from 'rxjs/Observable';
 
-import { AsyHttp, HttpOptions } from 'app/shared/asy-http.service';
-import { PagingOptions } from 'app/shared/pager.component';
+import { PagingOptions, AsyHttp, HttpOptions } from 'app/shared';
 
 export class CacheEntry {
 	date: Date;
