+++ src/client/app/admin/admin.component.ts
@@ -3,5 +3,4 @@ import { Component } from '@angular/core';
 @Component({
 	templateUrl: './admin.component.html'
 })
-export class AdminComponent {
-}
+export class AdminComponent {}
