+++ src/client/app/admin/authentication/inactive.component.ts
@@ -1,11 +1,12 @@
 import { Component } from '@angular/core';
+
 import { User } from '../user.class';
+
 import { AuthenticationService } from './authentication.service';
 
 @Component({
 	templateUrl: './inactive.component.html'
 })
-
 export class InactiveComponent {
 	user: User;
