+++ src/client/app/admin/authentication/invalid-certificate.component.ts
@@ -3,13 +3,4 @@ import { Component } from '@angular/core';
 @Component({
 	templateUrl: './invalid-certificate.component.html'
 })
-
-export class InvalidCertificateComponent {
-
-	constructor(
-	) {
-	}
-
-	ngOnInit() {
-	}
-}
+export class InvalidCertificateComponent {}
