+++ src/client/app/admin/authentication/no-access.component.ts
@@ -1,11 +1,12 @@
 import { Component } from '@angular/core';
+
 import { User } from '../user.class';
+
 import { AuthenticationService } from './authentication.service';
 
 @Component({
 	templateUrl: './no-access.component.html'
 })
-
 export class NoAccessComponent {
 	user: User;
