+++ src/client/app/admin/authentication/unauthorized.component.ts
@@ -5,12 +5,6 @@ import { User } from '../user.class';
 @Component({
 	templateUrl: './unauthorized.component.html'
 })
-
 export class UnauthorizedComponent {
 	user: User;
-
-	constructor(
-	) {
-
-	}
 }
