+++ src/client/app/admin/end-user-agreement/admin-eua.module.ts
@@ -7,12 +7,12 @@ import { BootstrapModalModule, Modal } from 'angular2-modal/plugins/bootstrap';
 import { Ng2BootstrapModule } from 'ng2-bootstrap';
 
 import { AdminService } from '../admin.service';
+import { AlertService, UtilModule } from 'app/shared';
+
 import { AdminListEuasComponent } from './admin-list-euas.component';
 import { AdminCreateEuaComponent } from './admin-create-eua.component';
 import { AdminUpdateEuaComponent } from './admin-edit-eua.component';
 import { UserEuaComponent } from './user-eua.component';
-import { UtilModule } from 'app/shared/util.module';
-import { AlertService } from 'app/shared/alert.service';
 import { EuaAudit } from './audit/eua-audit.component';
 
 @NgModule({
