+++ src/client/app/admin/end-user-agreement/index.ts
@@ -4,4 +4,6 @@ export * from './admin-edit-eua.component';
 export * from './user-eua.component';
 export * from './audit/eua-audit.component';
 
+export * from './eua.service';
+
 export * from './admin-eua.module';
