+++ src/client/app/admin/password/reset-password.component.ts
@@ -2,9 +2,8 @@ import { Component } from '@angular/core';
 import { Response } from '@angular/http';
 import { Router, ActivatedRoute, Params } from '@angular/router';
 
-import { AuthenticationService } from '../authentication/authentication.service';
-import { UserStateService } from '../authentication/user-state.service';
-import { ConfigService } from 'app/core/config.service';
+import { AuthenticationService, UserStateService } from '../authentication';
+import { ConfigService } from 'app/core';
 
 @Component({
 	templateUrl: './reset-password.component.html'
