+++ src/client/app/admin/user-management/admin-user.module.ts
@@ -5,6 +5,11 @@ import { RouterModule } from '@angular/router';
 
 import { Ng2BootstrapModule } from 'ng2-bootstrap';
 
+import { AlertService, UtilModule } from 'app/shared';
+import { CacheEntriesService } from 'app/access-checker';
+import { TeamsModule } from 'app/teams';
+
+import { UserAudit } from './audit/user-audit.component';
 import { AdminService } from '../admin.service';
 import { AdminCreateUserComponent } from './admin-create-user.component';
 import { AdminUpdateUserComponent } from './admin-edit-user.component';
@@ -13,11 +18,6 @@ import { ExportUsersModal } from './export-users.component';
 import { UpdateUserComponent } from './edit-user.component';
 import { UserSignupComponent } from './user-signup.component';
 import { UserService } from '../users.service';
-import { UtilModule } from 'app/shared/util.module';
-import { AlertService } from 'app/shared/alert.service';
-import { CacheEntriesService } from 'app/access-checker/cache-entries.service';
-import { UserAudit } from './audit/user-audit.component';
-import { TeamsModule } from 'app/teams/teams.module';
 
 @NgModule({
 	imports: [
