+++ src/client/app/admin/user-management/audit/user-audit.component.ts
@@ -1,10 +1,12 @@
-import { AuditObjectTypes } from 'app/audit/audit.classes';
-import { DefaultAudit } from 'app/audit/audit-object.component';
 import { Component } from '@angular/core';
+
+import { AuditObjectTypes } from 'app/audit/audit.classes';
+import { DefaultAudit } from 'app/audit';
 
 @Component({
 	selector: 'user',
 	templateUrl: './user-audit.component.html'
 })
 export class UserAudit extends DefaultAudit {}
+
 AuditObjectTypes.registerType('user', UserAudit);
