+++ src/client/app/admin/users.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 
+import { AsyHttp, HttpOptions, PagingOptions } from 'app/shared';
+
 import { User } from './user.class';
-import { AsyHttp, HttpOptions } from 'app/shared/asy-http.service';
-import { PagingOptions } from 'app/shared/pager.component';
 
 @Injectable()
 export class UserService {
